// src/enums/enums.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  SUPERADMIN = 'SUPERADMIN',
  SITEENGINEER = 'SITEENGINEER',
  LEADENGINEER = 'LEADENGINEER',
  CUSTOMERCARE = 'CUSTOMERCARE',
  STOREKEEPER = 'STOREKEEPER',
  ACCOUNTANT = 'ACCOUNTANT',
  MANAGER = 'MANAGER',
  AGENT = 'AGENT',
  CLIENT = 'CLIENT',
  HR = 'HR',
  STAFF = 'STAFF',
}
// ...
export enum Currency {
  NGN = 'NGN',
}

export enum Channel {
  EMAIL = 'EMAIL',
  SMS = 'SMS',
  WHATSAPP = 'WHATSAPP',
  INAPP = 'INAPP',
}

export enum TargetType {
  BILL = 'BILL',
  INCIDENT = 'INCIDENT',
  FUNDREQUEST = 'FUNDREQUEST',
}

export enum PropertyType {
  PLOT = 'PLOT',
  HOUSE = 'HOUSE',
}

export enum TransactionType {
  DEBIT = 'DEBIT',
  CREDIT = 'CREDIT',
  ADJUSTMENT = 'ADJUSTMENT',
}

export enum StateStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  AVAILABLE = 'AVAILABLE',
  UNAVAILABLE = 'UNAVAILABLE',
  SUSPENDED = 'SUSPENDED',
  BLOCKED = 'BLOCKED',
  TRASHED = 'TRASHED',
  UNSEEN = 'UNSEEN',
  SEEN = 'SEEN',
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  DECLINED = 'DECLINED',
  DELETED = 'DELETED',
  PAID = 'PAID',
  PARTIALLYPAID = 'PARTIALLYPAID',
  UNPAID = 'UNPAID',
  PAYING = 'PAYING',
  REPLIED = 'REPLIED',
  PURCHASED = 'PURCHASED',
  RELEASED = 'RELEASED',
  ALLOCATED = 'ALLOCATED',
  CANCELED = 'CANCELED',
  VALID = 'VALID',
  EXPIRED = 'EXPIRED',
  RESOLVED = 'RESOLVED',
  INPROGRESS = 'INPROGRESS',
  COMPLETED = 'COMPLETED',
  RESERVED = 'RESERVED',
  SOLD = 'SOLD',
  RESCHEDULED = 'RESCHEDULED',
  SETTLED = 'SETTLED',
  SUPPLIED = 'SUPPLIED',
  PARTIALLYSETTLED = 'PARTIALLYSETTLED',
  PENDINGSETTLEMENT = 'PENDINGSETTLEMENT',
  SUCCESS = 'SUCCESS',
  QUEUED = 'QUEUED',
  SENT = 'SENT',
  FAILED = 'FAILED',
  CLOSED = 'CLOSED',
  DELIVERED = 'DELIVERED',
  REJECTED = 'REJECTED',
  DUE = 'DUE',
  CANCELLED = 'CANCELLED',
  ACKNOWLEDGED = 'ACKNOWLEDGED',
  REVOKED = 'REVOKED',
  UPDATED = 'UPDATED',
  OFFERED = 'OFFERED',
}